Configure Pry.
Set Pry not to trap interrupts since we need to be in charge of that.
Also set Pry to use Readline (this is the default) and clean up any
history it adds.

There is still a strange issue where the first time you press up
after Pry is used, it is ignored.